[#25260] Set type of numpy tensors to float32 for Pet Dataset

+++ dl_framework_analyzer/datasets/pet_dataset.py
@@ -144,7 +144,7 @@ class PetDataset(Dataset):
             img = Image.open(sample)
             img = img.convert('RGB')
             img = img.resize((224, 224))
-            npimg = np.array(img) / 255.0
+            npimg = np.array(img, dtype=np.float32) / 255.0
             if self.standardize:
                 npimg = (npimg - self.mean) / self.std
             result.append(npimg)
